+++ contracts/interfaces/IERC5192.sol
@@ -1,0 +1,21 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.20;
+
+/// @title IERC5192
+/// @dev Interface for Minimal Soulbound NFTs
+/// See https://eips.ethereum.org/EIPS/eip-5192
+interface IERC5192 {
+    /// @notice Emitted when the locking status of an SBT is changed.
+    /// @param tokenId The identifier for a token.
+    event Locked(uint256 tokenId);
+
+    /// @notice Emitted when the unlocking status of an SBT is changed.
+    /// @param tokenId The identifier for a token.
+    event Unlocked(uint256 tokenId);
+
+    /// @notice Returns the locking status of an SBT
+    /// @dev SBTs assigned to zero address are considered invalid, and queries
+    /// about them do throw.
+    /// @param tokenId The identifier for an SBT.
+    function locked(uint256 tokenId) external view returns (bool);
+}